добавляет код для карты через API
