+++ index.html
@@ -4,6 +4,7 @@
   <meta charset="utf-8">
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
+  <link rel="stylesheet" href="/leaflet/leaflet.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кексобукинг</title>
 </head>
@@ -19,7 +20,9 @@
     <section class="map">
 
       <!-- Блок для вставки карты -->
-      <div class="map__canvas" id="map-canvas"></div>
+      <div class="map__canvas" id="map-canvas">
+<!--        style="width: 640px; height: 480px;"-->
+      </div>
 
       <!-- Фильтрация объявлений -->
       <div class="map__filters-container">
@@ -231,5 +234,6 @@
     </div>
   </template>
   <script src="js/main.js" type="module"></script>
+  <script src="/leaflet/leaflet.js"></script>
 </body>
 </html>
